# openapi.yaml
openapi: 3.0.3
info:
  title: Helix Health Portal API
  description: |
    RESTful API for the Helix Health Portal outpatient patient portal.
    Used as the sample application for the QA automation course.
  version: 1.0.0
  contact:
    name: Helix Health Portal
servers:
  - url: http://localhost:4000/api
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiSuccess:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid credentials"
...
    Patient:
      type: object
      properties:
        id:
          type: integer
          example: 1
        mrn:
          type: string
          example: MRN-TEST-00001
        firstName:
          type: string
          example: TEST_Patient
        lastName:
          type: string
          example: One
        dob:
          type: string
          format: date
          example: "1985-06-15"
        gender:
          type: string
          enum: [male, female, other, prefer_not_to_say]
        phone:
          type: string
          example: "555-123-4567"
        email:
          type: string
          format: email
...
    Prescription:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        prescriberId:
          type: integer
        drugName:
          type: string
          example: Metformin
        drugNdc:
          type: string
          nullable: true
        dosage:
          type: string
          example: 500mg
        frequency:
          type: string
          example: twice daily
        route:
          type: string
          example: oral
        quantity:
          type: integer
          example: 30
        refillsRemaining:
          type: integer
          example: 3
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [active, discontinued, completed, on_hold]
        isControlled:
          type: boolean
        scheduleClass:
          type: string
          nullable: true
          enum: [II, III, IV, V]
        pharmacyName:
          type: string
          nullable: true
        pharmacyPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        perPage:
          type: integer
        totalPages:
          type: integer

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /patients:
    get:
      tags: [Patients]
      summary: List all patients (admin, provider, nurse)
      operationId: listPatients
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search by name or MRN
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated patient list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Patient'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /patients/{patientId}/prescriptions:
    get:
      tags: [Prescriptions]
      summary: List prescriptions for a patient
      operationId: listPrescriptions
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, discontinued, completed, on_hold, all]
      responses:
        '200':
          description: Prescription list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Prescription'

    post:
      tags: [Prescriptions]
      summary: Create a prescription (provider/admin only)
      operationId: createPrescription
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [drug_name, dosage, frequency, start_date]
              properties:
                drug_name:
                  type: string
                  example: Metformin
                drug_ndc:
                  type: string
                  nullable: true
                dosage:
                  type: string
                  example: 500mg
                frequency:
                  type: string
                  example: twice daily
                route:
                  type: string
                  default: oral
                quantity:
                  type: integer
                  default: 30
                refills_remaining:
                  type: integer
                  default: 0
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                  nullable: true
                is_controlled:
                  type: integer
                  enum: [0, 1]
                  default: 0
                schedule_class:
                  type: string
                  nullable: true
                  enum: [II, III, IV, V]
                pharmacy_name:
                  type: string
                  nullable: true
                pharmacy_phone:
                  type: string
                  nullable: true
                notes:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Prescription created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Prescription'
                      warnings:
                        type: array
                        items:
                          type: string
                        description: Drug interaction warnings (if any)
        '400':
          description: Validation error (missing fields or dosage exceeds max)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Provider record not found or insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /refill-requests:
    get:
      tags: [Prescriptions]
      summary: List pending refill requests (provider view)
      operationId: listRefillRequests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending refill requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
  /refill-requests/{requestId}:
    put:
      tags: [Prescriptions]
      summary: Approve or deny a refill request
      operationId: updateRefillRequest
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, denied]
                notes:
                  type: string
      responses:
        '200':
          description: Refill request status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'

  # ── Drug Interactions ─────────────────────────────────────────────────────────
  /drug-interactions/check:
    post:
      tags: [Drug Interactions]
      summary: Check for interactions between a new drug and a patient's active medications
      operationId: checkDrugInteractions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientId, drugName]
              properties:
                patientId:
                  type: integer
                drugName:
                  type: string
                  example: Warfarin
      responses:
        '200':
          description: Interaction check result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          interactions:
                            type: array
                            items:
                              type: object
                              properties:
                                drugA:
                                  type: string
                                drugB:
                                  type: string
                                severity:
                                  type: string
                                  enum: [mild, moderate, severe]
                                description:
                                  type: string

  # ── Messages ──────────────────────────────────────────────────────────────────
...